Reorganize functions used by Pipeline.run
