+++ haystack/core/pipeline/pipeline.py
@@ -9,6 +9,7 @@ from warnings import warn
 from haystack import logging, tracing
 from haystack.core.component import Component
 from haystack.core.errors import PipelineMaxLoops, PipelineRuntimeError
+from haystack.core.pipeline.base import _dequeue_component, _enqueue_component
 from haystack.telemetry import pipeline_running
 
 from .base import PipelineBase, _add_missing_input_defaults, _is_lazy_variadic
@@ -22,22 +23,6 @@ class Pipeline(PipelineBase):
 
     Orchestrates component execution according to the execution graph, one after the other.
     """
-
-    def _component_has_enough_inputs_to_run(self, name: str, inputs: Dict[str, Dict[str, Any]]) -> bool:
-        """
-        Returns True if the Component has all the inputs it needs to run.
-
-        :param name: Name of the Component as defined in the Pipeline.
-        :param inputs: The current state of the inputs divided by Component name.
-
-        :return: Whether the Component can run or not.
-        """
-        instance: Component = self.graph.nodes[name]["instance"]
-        if name not in inputs:
-            return False
-        expected_inputs = instance.__haystack_input__._sockets_dict.keys()  # type: ignore
-        current_inputs = inputs[name].keys()
-        return expected_inputs == current_inputs
 
     def _run_component(self, name: str, inputs: Dict[str, Any]) -> Dict[str, Any]:
         """
@@ -94,9 +79,7 @@ class Pipeline(PipelineBase):
 
             return res
 
-    # TODO: We're ignoring these linting rules for the time being, after we properly optimize this function we'll
-    # remove the noqa
-    def run(  # noqa: C901, PLR0912, PLR0915 pylint: disable=too-many-branches,too-many-locals
+    def run(  # noqa: PLR0915
         self, data: Dict[str, Any], debug: bool = False, include_outputs_from: Optional[Set[str]] = None
     ) -> Dict[str, Any]:
         """
@@ -311,43 +294,3 @@ class Pipeline(PipelineBase):
                                 inner[k] = v
 
             return final_outputs
-
-
-def _enqueue_component(
-    component_pair: Tuple[str, Component],
-    to_run: List[Tuple[str, Component]],
-    waiting_for_input: List[Tuple[str, Component]],
-):
-    """
-    Append a Component in the queue of Components to run if not already in it.
-
-    Remove it from the waiting list if it's there.
-
-    :param component_pair: Tuple of Component name and instance
-    :param to_run: Queue of Components to run
-    :param waiting_for_input: Queue of Components waiting for input
-    """
-    if component_pair in waiting_for_input:
-        waiting_for_input.remove(component_pair)
-
-    if component_pair not in to_run:
-        to_run.append(component_pair)
-
-
-def _dequeue_component(
-    component_pair: Tuple[str, Component],
-    to_run: List[Tuple[str, Component]],
-    waiting_for_input: List[Tuple[str, Component]],
-):
-    """
-    Removes a Component both from the queue of Components to run and the waiting list.
-
-    :param component_pair: Tuple of Component name and instance
-    :param to_run: Queue of Components to run
-    :param waiting_for_input: Queue of Components waiting for input
-    """
-    if component_pair in waiting_for_input:
-        waiting_for_input.remove(component_pair)
-
-    if component_pair in to_run:
-        to_run.remove(component_pair)
