+++ test/core/pipeline/test_pipeline.py
@@ -15,8 +15,8 @@ from haystack.core.component import component
 from haystack.core.component.types import InputSocket, OutputSocket, Variadic
 from haystack.core.errors import PipelineConnectError, PipelineDrawingError, PipelineError
 from haystack.core.pipeline import Pipeline, PredefinedPipeline
-from haystack.core.pipeline.pipeline import _enqueue_component, _dequeue_component
-from haystack.core.pipeline.base import _add_missing_input_defaults
+from haystack.core.pipeline.base import _dequeue_component
+from haystack.core.pipeline.base import _add_missing_input_defaults, _enqueue_component
 from haystack.core.serialization import DeserializationCallbacks
 from haystack.testing.factory import component_class
 from haystack.testing.sample_components import AddFixedValue, Double, Greet
